requirements.txt
beautifulsoup4==4.8.2
lxml==4.5.0
python-dotenv==0.12.0
requests==2.23.0
